# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce API
  version: 0.2.0
  description: API contract for backend and frontend type generation.
servers:
  - url: http://localhost:3000
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /api/v1/checkout/orders/{id}/payments/authorize:
    post:
      tags: [checkout, orders]
      operationId: authorizeCheckoutOrderPayment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorizeCheckoutOrderPaymentRequest"
      responses:
        "200":
          description: Payment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessPaymentResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Guest checkout disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Checkout submission rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    Brand:
      type: object
      required:
        - id
        - name
        - slug
        - is_active
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        logo_media_id:
          type: string
          nullable: true
        is_active:
          type: boolean
...
    ProductOptionValue:
      type: object
      required:
        - value
        - position
      properties:
        id:
          type: integer
          minimum: 1
        value:
          type: string
        position:
          type: integer
          minimum: 1
    ProductOption:
      type: object
      required:
        - name
        - position
        - display_type
        - values
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
        position:
          type: integer
          minimum: 1
        display_type:
          type: string
        values:
          type: array
          items:
            $ref: "#/components/schemas/ProductOptionValue"
    ProductVariantSelection:
      type: object
      required:
        - option_name
        - option_value
        - position
      properties:
        product_option_value_id:
          type: integer
          minimum: 1
        option_name:
          type: string
        option_value:
          type: string
        position:
          type: integer
          minimum: 1
    ProductVariant:
      type: object
      required:
        - sku
        - title
        - price
        - stock
        - position
        - is_published
        - selections
      properties:
        id:
          type: integer
          minimum: 1
        sku:
          type: string
        title:
          type: string
        price:
          type: number
          format: double
        compare_at_price:
          type: number
          format: double
          nullable: true
        stock:
          type: integer
          minimum: 0
        position:
          type: integer
          minimum: 1
        is_published:
          type: boolean
        weight_grams:
          type: integer
          nullable: true
        length_cm:
          type: number
          format: double
          nullable: true
        width_cm:
          type: number
          format: double
          nullable: true
        height_cm:
          type: number
          format: double
          nullable: true
        selections:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariantSelection"
    ProductAttributeValue:
      type: object
      required:
        - product_attribute_id
        - key
        - slug
        - type
        - position
      properties:
        product_attribute_id:
          type: integer
          minimum: 1
        key:
          type: string
        slug:
          type: string
        type:
          type: string
        text_value:
          type: string
          nullable: true
        number_value:
          type: number
          format: double
          nullable: true
        boolean_value:
          type: boolean
          nullable: true
        enum_value:
          type: string
          nullable: true
        position:
          type: integer
          minimum: 1
    ProductSEO:
      type: object
      properties:
        title:
          type: string
          nullable: true
          maxLength: 70
        description:
          type: string
          nullable: true
          maxLength: 160
        canonical_path:
          type: string
          nullable: true
          maxLength: 255
        og_image_media_id:
          type: string
          nullable: true
        noindex:
          type: boolean
    ProductPriceRange:
      type: object
      required:
        - min
        - max
      properties:
        min:
          type: number
          format: double
        max:
          type: number
          format: double
    Product:
      type: object
      required:
        - id
        - sku
        - name
        - description
        - price
        - stock
        - images
        - related_products
        - options
        - variants
        - attributes
        - seo
        - price_range
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          minimum: 1
        sku:
          type: string
        name:
          type: string
        subtitle:
          type: string
          nullable: true
        description:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
          minimum: 0
        images:
          type: array
          items:
            type: string
        cover_image:
          type: string
          nullable: true
        brand:
          $ref: "#/components/schemas/Brand"
        default_variant_id:
          type: integer
          minimum: 1
          nullable: true
        default_variant_sku:
          type: string
          nullable: true
        price_range:
          $ref: "#/components/schemas/ProductPriceRange"
        options:
          type: array
          items:
            $ref: "#/components/schemas/ProductOption"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ProductAttributeValue"
        seo:
          $ref: "#/components/schemas/ProductSEO"
        related_products:
          type: array
          items:
            $ref: "#/components/schemas/RelatedProduct"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
        is_published:
          type: boolean
        has_draft_changes:
          type: boolean
        draft_updated_at:
          type: string
          format: date-time
          nullable: true
    RelatedProduct:
      type: object
      required:
        - id
        - sku
        - name
        - description
        - stock
      properties:
        id:
          type: integer
          minimum: 1
        sku:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: double
        stock:
          type: integer
          minimum: 0
        cover_image:
          type: string
          nullable: true
...
    OrderItem:
      type: object
      required:
        - id
        - order_id
        - product_variant_id
        - variant_sku
        - variant_title
        - quantity
        - price
        - product_variant
        - product
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        order_id:
          type: integer
        product_variant_id:
          type: integer
        variant_sku:
          type: string
        variant_title:
          type: string
        quantity:
          type: integer
        price:
          type: number
          format: double
        product_variant:
          $ref: "#/components/schemas/ProductVariant"
        product:
          $ref: "#/components/schemas/Product"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
    Order:
      type: object
      required:
        - id
        - checkout_session_id
        - status
        - can_cancel
        - total
        - items
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        user_id:
          type: integer
          nullable: true
        checkout_session_id:
          type: integer
        guest_email:
          type: string
          format: email
          nullable: true
        confirmation_token:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, PAID, FAILED, SHIPPED, DELIVERED, CANCELLED, REFUNDED]
        can_cancel:
          type: boolean
        total:
          type: number
          format: double
        payment_method_display:
          type: string
          nullable: true
        shipping_address_pretty:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
...
    AuthorizeCheckoutOrderPaymentRequest:
      type: object
      required: [snapshot_id]
      properties:
        snapshot_id:
          type: integer
          minimum: 1
    ProcessPaymentResponse:
      type: object
      required: [message, order]
      properties:
        message:
          type: string
        order:
          $ref: "#/components/schemas/Order"
...